+++ test/convert/metadataConverter.test.ts
@@ -5,21 +5,25 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import { createSandbox, SinonStub } from 'sinon';
-import { xmlInFolder, mockRegistry } from '../mock/registry';
+import { mockRegistry, xmlInFolder } from '../mock/registry';
 import * as streams from '../../src/convert/streams';
 import * as fs from 'graceful-fs';
 import * as fsUtil from '../../src/utils/fileSystemHandler';
 import { dirname, join } from 'path';
-import { expect, assert } from 'chai';
+import { assert, expect } from 'chai';
 import { ConversionError, LibraryError } from '../../src/errors';
 import { COMPONENTS } from '../mock/registry/type-constants/mixedContentInFolderConstants';
 import { fail } from 'assert';
-import { ComponentSet, MetadataConverter, SourceComponent } from '../../src';
+import {
+  ComponentSet,
+  DestructiveChangesType,
+  MetadataConverter,
+  SourceComponent,
+} from '../../src';
 import {
   DECOMPOSED_CHILD_COMPONENT_1,
   DECOMPOSED_CHILD_COMPONENT_2,
 } from '../mock/registry/type-constants/decomposedConstants';
-import { DestructiveChangesType } from '../../src/collections/types';
 
 const env = createSandbox();
 
@@ -160,7 +164,7 @@ describe('MetadataConverter', () => {
       ]);
     });
 
-    it('should write destructive changes post manifest when ComponentSet has deletes', async () => {
+    it('should write destructive changes post manifest when ComponentSet has deletes marked for post', async () => {
       const timestamp = 123456;
       const packagePath = join(
         outputDirectory,
@@ -172,14 +176,17 @@ describe('MetadataConverter', () => {
         type: DECOMPOSED_CHILD_COMPONENT_1.type,
         xml: DECOMPOSED_CHILD_COMPONENT_1.xml,
       });
-      component1.setMarkedForDelete(true);
+      component1.setMarkedForDelete(DestructiveChangesType.POST);
       const component2 = new SourceComponent({
         name: DECOMPOSED_CHILD_COMPONENT_2.name,
         type: DECOMPOSED_CHILD_COMPONENT_2.type,
         xml: DECOMPOSED_CHILD_COMPONENT_2.xml,
       });
       const compSet = new ComponentSet([component1, component2], mockRegistry);
-      const expectedDestructiveContents = compSet.getPackageXml(undefined, true);
+      const expectedDestructiveContents = compSet.getPackageXml(
+        undefined,
+        DestructiveChangesType.POST
+      );
       const expectedContents = compSet.getPackageXml();
 
       await converter.convert(compSet, 'metadata', { type: 'directory', outputDirectory });
@@ -207,7 +214,7 @@ describe('MetadataConverter', () => {
         type: DECOMPOSED_CHILD_COMPONENT_1.type,
         xml: DECOMPOSED_CHILD_COMPONENT_1.xml,
       });
-      component1.setMarkedForDelete(true);
+      component1.setMarkedForDelete(DestructiveChangesType.PRE);
       const component2 = new SourceComponent({
         name: DECOMPOSED_CHILD_COMPONENT_2.name,
         type: DECOMPOSED_CHILD_COMPONENT_2.type,
@@ -215,7 +222,10 @@ describe('MetadataConverter', () => {
       });
       const compSet = new ComponentSet([component1, component2], mockRegistry);
       compSet.setDestructiveChangesType(DestructiveChangesType.PRE);
-      const expectedDestructiveContents = compSet.getPackageXml(undefined, true);
+      const expectedDestructiveContents = compSet.getPackageXml(
+        undefined,
+        DestructiveChangesType.PRE
+      );
       const expectedContents = compSet.getPackageXml();
 
       await converter.convert(compSet, 'metadata', { type: 'directory', outputDirectory });
@@ -367,14 +377,17 @@ describe('MetadataConverter', () => {
         type: DECOMPOSED_CHILD_COMPONENT_1.type,
         xml: DECOMPOSED_CHILD_COMPONENT_1.xml,
       });
-      component1.setMarkedForDelete(true);
+      component1.setMarkedForDelete();
       const component2 = new SourceComponent({
         name: DECOMPOSED_CHILD_COMPONENT_2.name,
         type: DECOMPOSED_CHILD_COMPONENT_2.type,
         xml: DECOMPOSED_CHILD_COMPONENT_2.xml,
       });
       const compSet = new ComponentSet([component1, component2], mockRegistry);
-      const expectedDestructiveContents = compSet.getPackageXml(undefined, true);
+      const expectedDestructiveContents = compSet.getPackageXml(
+        undefined,
+        DestructiveChangesType.POST
+      );
       const expectedContents = compSet.getPackageXml();
       const addToZipStub = env.stub(streams.ZipWriter.prototype, 'addToZip');
 
@@ -391,13 +404,13 @@ describe('MetadataConverter', () => {
       ]);
     });
 
-    it('should write destructive changes pre manifest when ComponentSet has deletes', async () => {
+    it('should write destructive changes pre manifest when ComponentSet has deletes marked for pre', async () => {
       const component1 = new SourceComponent({
         name: DECOMPOSED_CHILD_COMPONENT_1.name,
         type: DECOMPOSED_CHILD_COMPONENT_1.type,
         xml: DECOMPOSED_CHILD_COMPONENT_1.xml,
       });
-      component1.setMarkedForDelete(true);
+      component1.setMarkedForDelete(DestructiveChangesType.PRE);
       const component2 = new SourceComponent({
         name: DECOMPOSED_CHILD_COMPONENT_2.name,
         type: DECOMPOSED_CHILD_COMPONENT_2.type,
@@ -405,7 +418,7 @@ describe('MetadataConverter', () => {
       });
       const compSet = new ComponentSet([component1, component2], mockRegistry);
       compSet.setDestructiveChangesType(DestructiveChangesType.PRE);
-      const expectedDestructiveContents = compSet.getPackageXml(undefined, true);
+      const expectedDestructiveContents = compSet.getPackageXml(4, DestructiveChangesType.PRE);
       const expectedContents = compSet.getPackageXml();
       const addToZipStub = env.stub(streams.ZipWriter.prototype, 'addToZip');